chore: refactored exportTemplate and moved logging into separate functions and file
